re-name the region layer and update the caffe
The new caffe code will print the debug info every 100 iterations by
default. The parameter of debug_info will be used as how many iteration
to print the debug info, instead of boolean. It should be enough to
remove it.

The region layer has a new parameter about the interations. If the layer
name is still 'loss', the model initialization from imagenet pretrained
model might fail since the name is the same but the numbers of parameters
are not. If we don't have this parameter, the seen_image will always be
0 at the beginning, no matter it is from scratch or frome a fine-tuned
model.